Commit to move data out of models
maintain(models/*, data/*): move data out of models
- create seperate data folder for data and files for user, account, transaction
- move the data out of models
- [Delivers #165381204]
